DA_BUBBLE/requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
fonttools==4.50.0
kiwisolver==1.4.5
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
